% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBIFsiteChecker-package.R
\docType{package}
\name{GBIFsiteChecker-package}
\alias{GBIFsiteChecker}
\alias{GBIFsiteChecker-package}
\title{GBIFsiteChecker}
\description{
GBIFsiteChecker
}
\details{
GBIFsiteChecker validates GBIF data according to their location.
}
\keyword{package}

