% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurence_access.r
\name{getNumberOfOccurences}
\alias{getNumberOfOccurences}
\title{Get the number of occurrences for a specific species}
\usage{
getNumberOfOccurences(species_name)
}
\arguments{
\item{species_name}{Name of the species to load}
}
\value{
species$meta$count - The number of occurrences for the specified 
  species
}
\description{
Get the number of occurrences for a specific species
}

