% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurence_access.r
\name{getSpeciesOpts}
\alias{getSpeciesOpts}
\title{Get the species properties from the worms database to check marine species
with failing location check}
\usage{
getSpeciesOpts(species_name)
}
\arguments{
\item{species_name}{Name of the species to get the properties for}
}
\value{
speciesOpts - habitat properties of the named species
}
\description{
Get the species properties from the worms database to check marine species
with failing location check
}

