% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurence_access.r
\name{initData}
\alias{initData}
\title{Loading data from the gbif database}
\usage{
initData(species_name, start = 0, limit = 200)
}
\arguments{
\item{species_name}{Specify the name of the species in order to download the
data from the gbif database}

\item{start}{Specify the start index to begin the download with}

\item{limit}{Specify the number of occurrences to download (max: 200.000)}
}
\description{
Loading data from the gbif database
}

