% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurence_access.r
\name{mainLoop}
\alias{mainLoop}
\title{Main loop to iterate over all occurences with variables chunks
(variable in size an number)}
\usage{
mainLoop(species_name, countries, correction_level, records_per_chunk,
  number_of_records, species_opts, progressbar = TRUE)
}
\arguments{
\item{species_name}{Name of the species to download the occurences}

\item{countries}{All countries of the world from shapefile to make
location checks.}

\item{correction_level}{A number defining the level of correction:
1 - flag (flag and dont correct invalid data),
2 - correct (flag and correct invalid data),
3 - strict (remove invalid data)}
}
\value{
correctedData - Checked and corrected location gbif data
}
\description{
Main loop to iterate over all occurences with variables chunks
(variable in size an number)
}

