% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurence_access.r
\name{pipeline}
\alias{pipeline}
\title{Feeding pipline (consinsting of the defined functions) with current_occ_chunk
and receiving the output result (with corrected or flagged output)}
\usage{
pipeline(current_occ_chunk)
}
\value{
current_occ_chunk_corrected: Corrected or unchanged but flagged 
  occurences
}
\description{
Feeding pipline (consinsting of the defined functions) with current_occ_chunk
and receiving the output result (with corrected or flagged output)
}

