% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurence_access.r
\name{plotPoint}
\alias{plotPoint}
\title{Plot one occurence based on the correction_flag}
\usage{
plotPoint(occ, countries)
}
\arguments{
\item{countries}{countries Dataset of all countries in the world provided by
natural earth.}

\item{current_occ}{Current occurence for plotting}
}
\description{
Plot one occurence based on the correction_flag
}

