% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurence_access.r
\name{startup}
\alias{startup}
\title{entry point of the package and decide for level of correction}
\usage{
startup(species_name, number_of_records, records_per_chunk,
  correction_level = 2)
}
\arguments{
\item{correction_level}{A number defining the level of correction:
1 - flag (flag and dont correct invalid data),
2 - correct (flag and correct invalid data),
3 - strict (remove invalid data)}
}
\description{
entry point of the package and decide for level of correction
}

